Place client and its factory in a single file
